refactored metrics form tag edit elements
introduced class TagsPanel to encapsulate the form elements and logic
for manipulating tags as a preparation to reuse the panel for the
annotations tags manipulation
